Major update: limit radix to 32-bit uns. integer
Define radix type in file src/core/type_defs.hpp
Define index type in file src/core/type_defs.hpp (to make sure that the
indexes that traverse the vector are unsigned 32-bit integer).
Parameter of operators in integer class are now int32_t.

Update dependencies in Makefile.

+++ src/utils/command.hpp
@@ -29,7 +29,7 @@ class command {
 		string op;
 		
 		integer big_value;
-		size_t small_value;
+		uint32_t small_value;
 		
 		list<command> sub_command;
 		vector<string> sorted_list;
